python/requirements.txt
RPi.GPIO==0.7.1
Flask==3.0.0
